added thumbnail logic for movies & tv shows
the TMDB API key is being used for generation.
the db (trends table) has been updated to include the image_urls for each image.
the api (api/generate-images) is being triggered by manually hitting the endpoint or when the image_url columns (for now only in tv-shows and movies row) are empty).